% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_matched_names}
\alias{check_matched_names}
\title{Check mount matched names}
\usage{
check_matched_names(tbl_mn, tbl_mmn)
}
\arguments{
\item{tbl_mn}{A data frame with the matched names.}

\item{tbl_mmn}{A data frame with the missmatched names.}
}
\value{
The check message formated and printed.
}
\description{
Ask if you want to show the matched names mounted.
}
