% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_stopwords}
\alias{check_stopwords}
\title{Check mount stopwords}
\usage{
check_stopwords(tbl)
}
\arguments{
\item{tbl}{A data frame with the stopwords.}
}
\value{
The check message formated and printed.
}
\description{
Ask if you want to show the stopwords mounted.
}
