% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_update_matched_names}
\alias{check_update_matched_names}
\title{Check update matched names}
\usage{
check_update_matched_names(tbl_mn_new, tbl_mn_old)
}
\arguments{
\item{tbl_mn_new}{A data frame with the new matching data.}

\item{tbl_mn_old}{A data frame with the old matching data.}
}
\value{
The check update message formated and printed.
}
\description{
Ask if you want to show the matched names updated.
}
