% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_update_stopwords}
\alias{check_update_stopwords}
\title{Check update stopwords}
\usage{
check_update_stopwords(tbl_sw_new, tbl_sw_old)
}
\arguments{
\item{tbl_sw_new}{A data frame with the new stopwords.}

\item{tbl_sw_old}{A data frame with the old stopwords.}
}
\value{
The check update message formated and printed.
}
\description{
Ask if you want to show the stopwords updated.
}
