% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_n_squish}
\alias{clear_n_squish}
\title{Clear string}
\usage{
clear_n_squish(string, my_stopwords)
}
\arguments{
\item{string}{A vector character.}

\item{my_stopwords}{A vector character with stopwords.}
}
\value{
A string corrected.
}
\description{
Removing stopwords, accents and squishing.
}
