% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_dsn}
\alias{connect_dsn}
\title{Connect to RDS Databases}
\usage{
connect_dsn(dsn = "Amazon RDS mercadoedu")
}
\arguments{
\item{dsn}{A string of mercadoedu data source name in ODBC.\strong{\preformatted{Default: "Amazon RDS mercadoedu"
}

}}
}
\value{
The connection of AWS RDS Database.
}
\description{
Use a data source name to connect on AWS RDS Database.
}
