% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct_n_lower}
\alias{correct_n_lower}
\title{Correct and lowercase}
\usage{
correct_n_lower(string)
}
\arguments{
\item{string}{A vector character.}
}
\value{
A string corrected and lowered.
}
\description{
Correct some characters errors with str_replace_all and
lower them.
}
