% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_objects_s3}
\alias{delete_objects_s3}
\title{Delete objects from S3}
\usage{
delete_objects_s3(bucket_objects, bucket_name = "price-quarentine")
}
\arguments{
\item{bucket_objects}{A vector character(name of objects in the AWS S3
bucket).}

\item{bucket_name}{A string of bucket name in AWS S3.\strong{\preformatted{Default: "price-quarentine"
}

}}
}
\description{
Delete objects from the AWS S3 bucket.
}
