% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{division}
\alias{division}
\title{Division function}
\usage{
division(...)
}
\arguments{
\item{...}{The values to be divided (will be concatenated with \code{c()}).}
}
\value{
The number resulting from the division.
}
\description{
Divide values.
}
