% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_matching_data}
\alias{get_matching_data}
\title{Get matching data}
\usage{
get_matching_data(
  con,
  tbl_name_ftk = "me_v3_pricing_fromtokey",
  tbl_name_pc = "pricing_course"
)
}
\arguments{
\item{con}{A database connection.}

\item{tbl_name_ftk}{A string with the name of the fromtokey table in
mercadoedu's database.\strong{\preformatted{Default: "me_v3_pricing_fromtokey"
}

}}

\item{tbl_name_pc}{A string with the name of the pricing_course table in
mercadoedu's database.\strong{\preformatted{Default: "pricing_course"
}

}}
}
\value{
The matching data in a data frame format.
}
\description{
Get matching data from "fromtokey" table in mercadoedu's
database(AWS RDS).
}
