% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_quarantine_data}
\alias{get_quarantine_data}
\title{Get quarantine data}
\usage{
get_quarantine_data(objects, bucket_name = "price-quarentine", dist = TRUE)
}
\arguments{
\item{objects}{A list of string(objects names in S3 bucket).}

\item{bucket_name}{A string of bucket name in S3.\strong{\preformatted{Default: "price-quarentine"
}

}}

\item{dist}{A logical bolean to apply \code{dplyr::distinct}.\strong{\preformatted{Default: "TRUE"
}

}}
}
\value{
The quarantine data in a data frame format.
}
\description{
Get quarantine data from csv files objects in AWS S3 bucket.
}
