% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_stopwords}
\alias{get_stopwords}
\title{Get stopwords}
\usage{
get_stopwords(con, tbl_name = "stopwords")
}
\arguments{
\item{con}{A database connection.}

\item{tbl_name}{A string with the name of the stopwords table in mercadoedu's
database.\strong{\preformatted{Default: "stopwords"
}

}}
}
\value{
The stopwords in a list.
}
\description{
Get stopwords from "stopwords"(default) table in
mercadoedu's database(AWS RDS).
}
