% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_color}
\alias{has_color}
\title{Has color}
\usage{
has_color(x)
}
\arguments{
\item{x}{A string with a hexadecimal color.}
}
\value{
A logical.
}
\description{
Check it out if it has the color(in colors list or in
hexadecimal).
}
