% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_hex_color}
\alias{has_hex_color}
\title{Has hex color}
\usage{
has_hex_color(color)
}
\arguments{
\item{x}{A string with a hexadecimal color.}
}
\value{
A logical.
}
\description{
Check if the string is a hexadecimal color(have length equals to
7 and starts with \verb{#}).
}
