% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_csv_s3}
\alias{import_csv_s3}
\title{Import CSV from S3}
\usage{
import_csv_s3(obj_csv, obj_n, bkt_name, n_max, id = NULL)
}
\arguments{
\item{obj_csv}{A vector character(name of object in S3 bucket).}

\item{obj_n}{A vector integer(index of object).}

\item{bkt_name}{A vector character(name of S3 bucket).}

\item{n_max}{A vector integer(total number of objects).}

\item{id}{A vector character(id of progress bar) to update.\strong{\preformatted{Default: "NULL"
}

}}
}
\value{
A S3 file object csv in a data frame format.
}
\description{
Import CSV file object from S3 bucket with vroom package.
}
