% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_join}
\alias{make_join}
\title{Make join}
\usage{
make_join(tbl_nn, tbl_db, by_n, n_iter)
}
\arguments{
\item{tbl_nn}{A data frame with new names of the courses.}

\item{tbl_db}{A data frame with the names of the courses.}

\item{by_n}{A list of variables combination to join.}

\item{n_iter}{A number of iterations.}
}
\value{
A data frame joined by \code{by_n}.
}
\description{
A function to make joins.
}
