% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_joineds_append}
\alias{make_joineds_append}
\title{Make Joineds Tables Append}
\usage{
make_joineds_append(tbl_matching, tbl_new_names)
}
\arguments{
\item{tbl_matching}{A data frame with the names of the courses.}

\item{tbl_new_names}{A data frame with new names of the courses.}
}
\value{
A data frame with all joins for course names.
}
\description{
A function to make append of joineds tables.
}
