% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mount.R
\name{mount_matched_names}
\alias{mount_matched_names}
\title{Mount matched names}
\usage{
mount_matched_names(
  con,
  tbl_name_ftk = "me_v3_pricing_fromtokey",
  tbl_name_pc = "pricing_course",
  tbl_name_sw = "stopwords",
  bucket_name = "price-quarentine",
  objects_prefix = "run-AmazonS3_node",
  check = TRUE
)
}
\arguments{
\item{con}{A PostgreSQL's ODBC connection to Amazon RDS mercadoedu.}

\item{tbl_name_ftk}{A string of table name in mercadoedu database.\strong{\preformatted{Default: "me_v3_pricing_fromtokey"
}

}}

\item{tbl_name_pc}{A string of table name in mercadoedu database.\strong{\preformatted{Default: "pricing_course"
}

}}

\item{tbl_name_sw}{A string of stopwords table name in mercadoedu database.
\strong{\preformatted{Default: "stopwords"
}

}}

\item{bucket_name}{A string of bucket name in AWS S3.\strong{\preformatted{Default: "price-quarentine"
}

}}

\item{objects_prefix}{A string of prefix to filter objects from bucket.
\strong{\preformatted{Default: "run-AmazonS3_node"
}

}}

\item{check}{A logical that defines if check results or not(asking to show).
\strong{\preformatted{Default: TRUE
}

}}
}
\value{
List with \code{tables}(\code{matched} and \code{missmatched} in data frame format)
and bucket_objects.
}
\description{
Mount course names matcheds in a data frame format.
}
