% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mount.R
\name{mount_stopwords}
\alias{mount_stopwords}
\title{Mount stopwords data}
\usage{
mount_stopwords(sw_list, sw_wrongs_list, check = TRUE)
}
\arguments{
\item{sw_list}{A list of stopwords as string vector.}

\item{sw_wrongs_list}{A list of wrongs stopwords as string vector.}

\item{check}{A logical that defines if check results or not(asking to show).
\strong{\preformatted{Default: TRUE
}

}}
}
\value{
A stopwords table in data frame format.
}
\description{
Mount stopwords table in a data frame format.
}
