% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{msg_error}
\alias{msg_error}
\title{Message Error}
\usage{
msg_error(text, color = "red", symbol = NULL)
}
\arguments{
\item{text}{A vector character(reference in string format to be paste).}

\item{color}{A vector character(a color).\strong{\preformatted{Default: "red"
}

}}

\item{symbol}{A vector character(a symbol e.g.: "*").\strong{\preformatted{Default: "NULL"
}

}}
}
\value{
The text named with \code{symbol} icon.
}
\description{
Formated text for errors.
}
