% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{msg_ref}
\alias{msg_ref}
\title{Message Reference}
\usage{
msg_ref(text)
}
\arguments{
\item{text}{A vector character(reference in string format to be paste).}
}
\value{
A italic and quoted reference in terminal.
}
\description{
Customized message for references.
}
