% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiplication}
\alias{multiplication}
\title{Multiplication function}
\usage{
multiplication(...)
}
\arguments{
\item{...}{The values to be multiplied (will be concatenated with \code{c()}).}
}
\value{
The number resulting from the multiplication.
}
\description{
Multiplies values.
}
