% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline_stopwords}
\alias{pipeline_stopwords}
\title{Pipeline stopwords}
\usage{
pipeline_stopwords(
  data_source_name,
  stopwords_list,
  stopwords_wrongs_list,
  table_name_stopwords,
  check = TRUE
)
}
\arguments{
\item{data_source_name}{A PostgreSQL's ODBC connection.}

\item{stopwords_list}{A list of stopwords as string vector.}

\item{stopwords_wrongs_list}{A list of wrongs stopwords as string vector.}

\item{table_name_stopwords}{A string of table name in mercadoedu's database.
\strong{\preformatted{Default: "stopwords"
}

}}

\item{check}{A logical that defines if check results or not(asking to show).
\strong{\preformatted{Default: TRUE
}

}}
}
\value{
List with \code{table}(stopwords table) and \code{update}(result of stopwords
update).
}
\description{
Connect to ODBC's data source, mount stopwords table and update
them.
}
