% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep_matching_data}
\alias{prep_matching_data}
\title{Prepare matching data frame}
\usage{
prep_matching_data(tbl_matching, my_stopwords)
}
\arguments{
\item{tbl_matching}{A data frame with matching data.}

\item{my_stopwords}{A list of stopwords.}
}
\value{
The matching data prepared in a data frame format.
}
\description{
Prepare matching data frame, correcting wrong characters,
misspelling and squishing names.
}
