% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep_quarantine_data}
\alias{prep_quarantine_data}
\title{Prepare quarantine data frame}
\usage{
prep_quarantine_data(tbl_quarantine, my_stopwords)
}
\arguments{
\item{tbl_quarantine}{A data frame with quarantine data.}

\item{my_stopwords}{A list of stopwords.}
}
\value{
The quarantine data prepared in a data frame format.
}
\description{
Prepare quarantine data frame, correcting wrong characters,
misspelling and squishing names.
}
