% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{put_csv_s3}
\alias{put_csv_s3}
\title{Put CSV in S3}
\usage{
put_csv_s3(tbl_mmn, bucket_name = "price-quarentine")
}
\arguments{
\item{tbl_mmn}{A data frame with the missmatched names.}

\item{bucket_name}{A string of bucket name in AWS S3.\strong{\preformatted{Default: "price-quarentine"
}

}}
}
\description{
Write a csv file from table and put into the AWS S3 bucket.
}
