% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_objects_s3}
\alias{show_objects_s3}
\title{Show objects from S3}
\usage{
show_objects_s3(bucket_name = "price-quarentine", prefixer = NULL)
}
\arguments{
\item{bucket_name}{A string(name of AWS S3 Bucket).\preformatted{Default: "price-quarentine"
}}

\item{prefixer}{A string(prefix of bucket's objects).\preformatted{Default: NULL
}}
}
\value{
A list of objects inside a bucket.
}
\description{
Get objects names inside a AWS S3 Bucket.
}
