% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subtraction}
\alias{subtraction}
\title{Subtraction function}
\usage{
subtraction(...)
}
\arguments{
\item{...}{The values to be subtracted (will be concatenated with \code{c()}).}
}
\value{
The number resulting from the subtraction.
}
\description{
Subtract values.
}
