% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_function}
\alias{ui_function}
\title{UI Function}
\usage{
ui_function(text, color = "#10aff2", type = "start", bl = "both")
}
\arguments{
\item{text}{A vector character (will be printed).}

\item{color}{A vector character (a color name or a hex color).
\strong{\preformatted{Default: "#10aff2"
}

}}

\item{type}{A vector character with the type of message ("start" or "end").
\strong{\preformatted{Default: "start"
}

}}

\item{bl}{A vector character with position of blank line
("after", "before", "both" or "none").
\strong{\preformatted{Default: "both"
}

}}
}
\value{
The function message formated and printed.
}
\description{
Print start/end function message.
}
