% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_info}
\alias{ui_info}
\title{UI Info}
\usage{
ui_info(text, bl = "after")
}
\arguments{
\item{text}{A vector character(reference in string format to be paste).}

\item{bl}{A vector character with position of blank line
("after", "before", "both" or "none").
\strong{\preformatted{Default: "both"
}

}}
}
\value{
A info message formated in terminal.
}
\description{
Customized message for references.
}
