% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_match}
\alias{ui_match}
\title{UI match}
\usage{
ui_match(id, name, db)
}
\arguments{
\item{id}{The disc_id value (must be a numeric).}

\item{name}{A vector character with the label of discr_id (must be a string).}

\item{db}{A database with the matched names.}
}
\value{
The results as a message formated and printed.
}
\description{
Print results by main name.
}
