% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_missmatch}
\alias{ui_missmatch}
\title{UI missmatch}
\usage{
ui_missmatch(original_name)
}
\arguments{
\item{original_name}{A vector character with original_name (must be string).}
}
\value{
The results as a message formated and printed.
}
\description{
Print results by main name.
}
