% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_percent}
\alias{ui_percent}
\title{UI Percent}
\usage{
ui_percent(percent, type, bl = "none")
}
\arguments{
\item{percent}{The value (must be numeric).}

\item{type}{A vector character with the type of message ("rows" or "names").}

\item{bl}{A vector character with position of blank line
("after", "before", "both" or "none").
\strong{\preformatted{Default: "both"
}

}}
}
\value{
The message of percentage formated and printed.
}
\description{
Print a message according to the percent and type.
}
