% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_success}
\alias{ui_success}
\title{UI Success}
\usage{
ui_success(text, bl = "none")
}
\arguments{
\item{text}{A vector character(reference in string format to be paste).}

\item{bl}{A vector character with position of blank line
("after", "before", "both" or "none").
\strong{\preformatted{Default: "both"
}

}}
}
\value{
A success message with icon in terminal.
}
\description{
Customized message for success.
}
