% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_matched_names}
\alias{update_matched_names}
\title{Update matching data table}
\usage{
update_matched_names(
  con,
  tbl_mn,
  tbl_name_ftk = "me_v3_pricing_fromtokey",
  tbl_name_pc = "pricing_course",
  check = TRUE
)
}
\arguments{
\item{con}{A PostgreSQL ODBC connection to Amazon RDS mercadoedu's database.}

\item{tbl_mn}{A data frame with matched names to be appended.}

\item{tbl_name_ftk}{A string of \verb{fromtokey table} name in mercadoedu's
database.\strong{\preformatted{Default: "me_v3_pricing_fromtokey"
}

}}

\item{tbl_name_pc}{A string of \verb{pricing_course table} name in mercadoedu's
database.\strong{\preformatted{Default: "pricing_course"
}

}}

\item{check}{A logical that defines when to check matched names.
\strong{\preformatted{Default: TRUE
}

}}
}
\value{
A logical that warnings if new course names was append or not.
}
\description{
Update a matching data table from database with new matches.
}
