% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_stopwords}
\alias{update_stopwords}
\title{Update stopwords table}
\usage{
update_stopwords(con, tbl_sw, tbl_name = "stopwords", check = TRUE)
}
\arguments{
\item{con}{A PostgreSQL ODBC connection to Amazon RDS mercadoedu's database.}

\item{tbl_sw}{A data frame with stopwords to be appended or overwritten.}

\item{tbl_name}{A string of table name in mercadoedu's database.\strong{\preformatted{Default: "stopwords"
}

}}

\item{check}{A logical that defines when to check stopwords.
\strong{\preformatted{Default: TRUE
}

}}
}
\value{
A logical that warnings if stopwords database was updated or not.
}
\description{
Update a stopwords table from database with new stopwords.
}
