% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_bl}
\alias{valid_bl}
\title{Valid blank line option}
\usage{
valid_bl(bl, bl_options)
}
\arguments{
\item{bl}{A string.}

\item{bl_options}{A vector character with the options.}
}
\value{
A logical.
}
\description{
Check it out if it is a valid blank line option(if is a valid
string or if are in the options).
}
