% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_color}
\alias{valid_color}
\title{Valid color}
\usage{
valid_color(x)
}
\arguments{
\item{x}{A string with a color name or a hexadecimal color(with or without
hashtag).}
}
\value{
A logical.
}
\description{
Check it out if it is a valid color(it has the color list, or it
is a hexadecimal color).
}
