% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_percent}
\alias{valid_percent}
\title{Valid percent}
\usage{
valid_percent(x)
}
\arguments{
\item{x}{A number.}
}
\value{
A logical.
}
\description{
Check it out if it is a valid percentage(not null, is numeric
and be between 0 and 100).
}
