% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_string}
\alias{valid_string}
\title{Valid string}
\usage{
valid_string(x)
}
\arguments{
\item{x}{A string.}
}
\value{
A logical.
}
\description{
Check it out if it is a valid string(not null, is character and
the length is more than 1).
}
