% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline_matched_names}
\alias{pipeline_matched_names}
\title{Pipeline matched names}
\usage{
pipeline_matched_names(
  data_source_name,
  table_name_from_to_key,
  table_name_pricing_course,
  table_name_stopwords,
  bucket_name,
  objects_prefix,
  check = TRUE
)
}
\arguments{
\item{bucket_name}{A string of bucket name in AWS S3.\strong{\preformatted{Default: "price-quarentine"
}

}}

\item{objects_prefix}{A string of prefix to filter objects from bucket.
\strong{\preformatted{Default: "run-AmazonS3_node"
}

}}

\item{check}{A logical that defines if check results or not(asking to show).
\strong{\preformatted{Default: TRUE
}

}}

\item{con}{A PostgreSQL's ODBC connection to Amazon RDS mercadoedu.}

\item{tbl_name_ftk}{A string of table name in mercadoedu database.\strong{\preformatted{Default: "me_v3_pricing_fromtokey"
}

}}

\item{tbl_name_pc}{A string of table name in mercadoedu database.\strong{\preformatted{Default: "pricing_course"
}

}}

\item{tbl_name_sw}{A string of stopwords table name in mercadoedu database.
\strong{\preformatted{Default: "stopwords"
}

}}
}
\value{
List with \code{tables}(\code{matched_names} and \code{missmatched_names}) and
\code{results}(results of \code{update}, \code{put} and \code{delete}).
}
\description{
Connect to ODBC's data source, mount matched names table, update
them and .
}
